requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
Pillow==11.0.0
python-decouple==3.8
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==4.0.0
requests==2.31.0
